+++ application/controllers/Pelanggan.php
@@ -37,6 +37,8 @@ class Pelanggan extends CI_Controller
     $data = [];
     $no = 1;
 
+    ($this->session->userdata('role_admin') == 0) ? $display = 'hidden-guest' : $display = ''; 
+
     foreach ($this->pelanggan->getAll() as $row) {
       $data[] = array(
         $no++,
@@ -44,9 +46,11 @@ class Pelanggan extends CI_Controller
         $row->email,
         $row->telpon,
         $row->status,
-        '<a href="' . base_url('pelanggan/view/' . $row->id_pelanggan) . '" class="btn btn-primary"><i class="fa fa-eye"></i></a>
-        <a href="' . base_url('pelanggan/edit/' . $row->id_pelanggan) . '" class="btn btn-warning"><i class="fa fa-edit"></i></a>
-        <button class="delete-button btn btn-danger" row-data="pelanggan-' . $row->id_pelanggan . '" data-url="' . base_url('pelanggan/delete/' . $row->id_pelanggan) . '">
+        '<button type="button" class="btn btn-primary detail-button" data-toggle="modal" data-target="#modal-detail" data-id="' . $row->id_pelanggan . '">
+          <i class="fa fa-eye"></i>
+        </button>
+        <a href="' . base_url('pelanggan/edit/' . $row->id_pelanggan) . '" class="btn btn-warning '.$display.'"><i class="fa fa-edit"></i></a>
+        <button class="delete-button btn btn-danger '.$display.'" row-data="pelanggan-' . $row->id_pelanggan . '" data-url="' . base_url('pelanggan/delete/' . $row->id_pelanggan) . '">
             <i class="fa fa-trash"></i>
         </button>'
       );
@@ -62,6 +66,11 @@ class Pelanggan extends CI_Controller
 
   public function store()
   {
+    if($this->session->userdata('role_admin') == 0){
+      $this->session->set_flashdata('error', '<i class="fa fa-exclamation-circle"></i>&nbsp; Access denied for guest!');
+      redirect(base_url('pelanggan'));
+    }
+
     $data = [
       'nama_pelanggan' => $this->input->post('nama_pelanggan'),
       'alamat' => $this->input->post('alamat'),
@@ -89,22 +98,20 @@ class Pelanggan extends CI_Controller
     redirect(base_url('pelanggan'));
   }
 
-  public function view($id)
+  public function detail()
   {
-    $this->db->where('id_pelanggan', $id);
-    $pelanggan = $this->db->get('pelanggan')->row();
-    $data = [
-      'title' => 'Lihat Pelanggan',
-      'page' => 'pelanggan/form_view',
-      'data' => $pelanggan,
-      'kategori' => $this->kategori->getAll(),
-      'user' => $this->user->getAll()
-    ];
-    $this->load->view('index', $data);
+    $id = $this->input->post('id');
+    header('Content-Type: application/json');
+    echo json_encode($this->pelanggan->getById($id));
   }
 
   public function create()
   {
+    if($this->session->userdata('role_admin') == 0){
+      $this->session->set_flashdata('error', '<i class="fa fa-exclamation-circle"></i>&nbsp; Access denied for guest!');
+      redirect(base_url('pelanggan'));
+    }
+
     $data = [
       'title' => 'Tambah Pelanggan',
       'page' => 'pelanggan/form_tambah',
@@ -116,6 +123,11 @@ class Pelanggan extends CI_Controller
 
   public function edit($id)
   {
+    if($this->session->userdata('role_admin') == 0){
+      $this->session->set_flashdata('error', '<i class="fa fa-exclamation-circle"></i>&nbsp; Access denied for guest!');
+      redirect(base_url('pelanggan'));
+    }
+
     $data = [
       'title' => 'Update Pelanggan',
       'page' => 'pelanggan/form_update',
@@ -127,6 +139,11 @@ class Pelanggan extends CI_Controller
 
   function update($id)
   {
+
+    if($this->session->userdata('role_admin') == 0){
+      $this->session->set_flashdata('error', '<i class="fa fa-exclamation-circle"></i>&nbsp; Access denied for guest!');
+      redirect(base_url('pelanggan'));
+    }
 
     $data = [
       'nama_pelanggan' => $this->input->post('nama_pelanggan'),
@@ -157,6 +174,11 @@ class Pelanggan extends CI_Controller
 
   public function delete($id)
   {
+    if($this->session->userdata('role_admin') == 0){
+      $this->session->set_flashdata('error', '<i class="fa fa-exclamation-circle"></i>&nbsp; Access denied for guest!');
+      redirect(base_url('pelanggan'));
+    }
+
     $pelanggan = $this->pelanggan->getById($id);
     $this->pelanggan->delete($id);
